move get unique ids to init
